TASK: Collect unit and end-to-end tests in tests/ConnectorTest.php
Reduce the structure of the test files to eliminate redundancies.
Currently there are only 3 tests, so 3 test files and two test
subfolders are a bit overkill :)
